requirements.txt
python-dotenv==1.0.1
transformers==4.41.0
accelerate==0.30.1
sentencepiece==0.2.0
pandas==2.0.3
openai==1.30.1
google-generativeai==0.5.4